+++ src/components/Buttons/Button/__tests__/Button.spec.tsx
@@ -5,6 +5,11 @@ import userEvent from '@testing-library/user-event';
 import { Button } from '../Button';
 
 describe('<Button> that renders a button', () => {
+    it('should not render any HTML if no children are provided', () => {
+        const view = render(<Button>{null}</Button>);
+
+        expect(view.container).toBeEmptyDOMElement();
+    });
     it('should render default button correctly', () => {
         const view = render(<Button>Click me</Button>);
 
@@ -37,7 +42,7 @@ describe('<Button> that renders a button', () => {
     it('should add string html attributes correctly', () => {
         render(<Button data-test="something">Click me</Button>);
 
-        expect(screen.getByRole('button').outerHTML).toMatch('data-test="something"');
+        expect(screen.getByRole('button')).toHaveAttribute('data-test', 'something');
     });
     it('should add functional html attributes correctly', async () => {
         const onMouseOverMock = jest.fn();