+++ src/components/Link/__tests__/Link.spec.js
@@ -3,6 +3,10 @@ import toJson from 'enzyme-to-json';
 import { Link } from '../Link';
 
 describe('<Link> that renders a link', () => {
+    it('should render nothing if no children are provided', () => {
+        const wrapper = shallow(<Link href="https://textkernel.com" />);
+        expect(wrapper.html()).toBeNull();
+    });
     it('should render default link correctly', () => {
         const wrapper = shallow(<Link href="https://textkernel.com">Click me</Link>);
         expect(toJson(wrapper)).toMatchSnapshot();