+++ src/components/WidthRestrictor/BlockWidthRestrictor/BlockWidthRestrictor.tsx
@@ -1,11 +1,10 @@
 import React, { forwardRef } from 'react';
-import { NotEmptyReactNode } from '../../../customTypes/types';
 import { bem } from '../../../utils';
 import styles from './BlockWidthRestrictor.scss';
 
 export interface Props extends React.HTMLAttributes<HTMLElement> {
     /** Node(s) to be rendered inside the container */
-    children: NotEmptyReactNode;
+    children: React.ReactNode;
     /** HTML tag to be used to render the container */
     As?: string;
 }
@@ -14,6 +13,10 @@ const { block } = bem('BlockWidthRestrictor', styles);
 
 export const BlockWidthRestrictor = forwardRef<HTMLElement, Props>(
     ({ children, As = 'div', ...rest }, ref) => {
+        if (!children) {
+            return null;
+        }
+
         return (
             <As {...rest} ref={ref} {...block(rest)}>
                 {children}